bugfix(produtos): Remoção em Cascada

+++ backend/src/StockSolution.Api/Features/Suppliers/DeleteSuppliers.cs
@@ -38,6 +38,12 @@ public sealed class DeleteSuppliersCommandHandler : IRequestHandler<DeleteSuppli
 
     public async Task Handle(DeleteSuppliers req, CancellationToken ct)
     {
+        var produtosComFornecedor = await _context.Products.AnyAsync(p => p.SupplierId == req.Id, ct);
+        if (produtosComFornecedor)
+        {
+            throw new Exception("Fornecedor Associado a um Produto. Não é Possível Removê-lo");
+        }
+
         await _context.Suppliers.Where(c => c.Id == req.Id).ExecuteDeleteAsync(ct);
 
         return;
